Add integration tests for Darwin providers
Also moved the provider out of the interfact in to its own class file.
